sla-oncall/requirements.txt
requests
environs
mysql-connector-python